refactor: 모달 스토리북 및 컴포넌트 수정
