+++ components/src/lib/Modal.stories.tsx
@@ -1,7 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import Modal from "./Modal";
-import { useState } from "react";
-import { Content, CloseContent } from "../App";
 import "../App.css";
 import { fn } from "@storybook/test";
 
@@ -21,6 +19,10 @@ const meta = {
       control: "object",
       description: "모달 컨테이너 인라인 스타일",
     },
+    position: {
+      options: ["center", "bottom"],
+      control: { type: "radio" },
+    },
   },
   args: {
     onClose: fn(),
@@ -31,38 +33,16 @@ export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Default: Story = {
+export const Center: Story = {
   args: {
     isOpen: true,
+    position: "center",
   },
-  render: (args) => {
-    const [modalOpen, setModalOpen] = useState(false);
+};
 
-    const handleModalClose = () => {
-      setModalOpen(false);
-    };
-
-    const handleModalOpen = () => {
-      setModalOpen(true);
-    };
-
-    return (
-      <>
-        <Modal isOpen={modalOpen} onClose={handleModalClose}>
-          <Modal.Portal id="modal">
-            <Modal.Backdrop>
-              <Modal.Container position="bottom" className="container">
-                <Content />
-                <Modal.ModalCloseButton>
-                  <CloseContent />
-                </Modal.ModalCloseButton>
-              </Modal.Container>
-            </Modal.Backdrop>
-          </Modal.Portal>
-        </Modal>
-
-        <button onClick={handleModalOpen}>모달 열기</button>
-      </>
-    );
+export const Bottom: Story = {
+  args: {
+    isOpen: true,
+    position: "bottom",
   },
 };
